+++ measurement/mpstat.py
@@ -12,13 +12,19 @@ class SysStat(measurement.Measurement):
         measurement.Measurement.__init__(self, experiment)
         self.process = None
 
-        subprocess.check_call(self.check_cmd.split(),
-                              stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-
-        self.live = utils.live.LiveStream()
-        self.headers = None
+        try:
+            subprocess.check_call(self.check_cmd.split(),
+                                  stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+        except Exception as e:
+            print(f"WARNING: {self.__class__.__name__}: failed to load ({self.check_cmd}: {e}")
+            self.live = False
+        else:
+            self.live = utils.live.LiveStream()
+            self.headers = None
 
     def start(self):
+        if not self.live: return
+
         def preexec(): # Don't forward signals.
             os.setpgrp()
 
@@ -28,6 +34,8 @@ class SysStat(measurement.Measurement):
         self.live.start(self.process.stdout)
 
     def stop(self):
+        if not self.live: return
+
         self.live.stop()
         self.process.kill()
